+++ src/test/java/PageObject/NavigateToRatePage.java
@@ -1,4 +1,4 @@
-package TestData;
+package PageObject;
 
 import org.openqa.selenium.*;
 import org.openqa.selenium.support.ui.ExpectedConditions;
