+++ src/test/java/PageObject/PaymentTerms.java
@@ -1,4 +1,4 @@
-package Start;
+package PageObject;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
